Move mlir tests
Move test_reshape_accuracy to mlir tests
Remove pytest param from embedding tests
