+++ forge/test/operators/pytorch/tm/test_reshape.py
@@ -308,27 +308,3 @@ TestParamsData.test_plan = TestPlan(
 
 def get_test_plans() -> List[TestPlan]:
     return [TestParamsData.test_plan]
-
-
-@pytest.mark.parametrize(
-    "input_shape, target_shape",
-    [
-        ((1, 4), (4, 1, 1)),
-        ((1, 100), (10, 10)),
-        ((1, 3, 4), (6, 2)),
-    ],
-)
-def test_reshape_accuracy(input_shape, target_shape):
-
-    inputs = [torch.ones(input_shape) - 0.1]  # value: 0.9000
-
-    pytorch_model = ModelDirect(torch.reshape, {"shape": target_shape})
-    compiled_model = forge.compile(pytorch_model, sample_inputs=inputs)
-
-    # forge.verify.verify.verify as forge_verify:
-    forge_verify(
-        inputs,
-        pytorch_model,
-        compiled_model,
-        verify_cfg=VerifyConfig(value_checker=AllCloseValueChecker()),
-    )
